PocketBeagle-2: rgb_led: hue: Add python example
